a.sh

+++ sh/a.sh
@@ -1,1 +1,5 @@
-rsync -ah --info=progress2 --delete --inplace --no-whole-file "$@"
+rsync -ahnv --delete "$@"
+read -p 'are you sure? ' sure
+if [ "$sure" = '' -o "$sure" = 'y' ]; then
+  rsync -ah --info=progress2 --delete --inplace --no-whole-file "$@"
+fi
